added sketches page to router and moved profile panel to common page
